+++ site/src/SoundfontExample.tsx
@@ -58,7 +58,7 @@ export function SoundfontExample({ className }: { className?: string }) {
       >
         <div className="flex gap-4 mb-2">
           <select
-            className="bg-zinc-700 rounded"
+            className="appearance-none bg-zinc-700 text-zinc-200 rounded border border-gray-400 py-2 px-3 leading-tight focus:outline-none focus:border-blue-500"
             value={libraryName}
             onChange={(e) => {
               const libraryName = e.target.value;
@@ -73,7 +73,7 @@ export function SoundfontExample({ className }: { className?: string }) {
             ))}
           </select>
           <select
-            className="bg-zinc-700 rounded"
+            className="appearance-none bg-zinc-700 text-zinc-200 rounded border border-gray-400 py-2 px-3 leading-tight focus:outline-none focus:border-blue-500"
             value={instrumentName}
             onChange={(e) => {
               const instrumentName = e.target.value;
@@ -132,7 +132,7 @@ export function SoundfontExample({ className }: { className?: string }) {
             instrument.start(note);
           }}
           onRelease={(midi) => {
-            instrument?.stop({ stopId: "" + midi });
+            instrument?.stop({ stopId: midi });
           }}
         />
       </div>