+++ src/electric-piano.ts
@@ -1,4 +1,4 @@
-import { midiVelToGain } from "./sampler/note";
+import { midiVelToGain } from "./sampler/midi";
 import { createControl } from "./sampler/signals";
 import { SfzSampler, SfzSamplerConfig } from "./sfz/sfz-sampler";
 import { createTremolo } from "./tremolo";